fix: changed support for double to int

+++ HWs/HW2/StackCalculator/StackCalculator/Stack/IStack.cs
@@ -9,14 +9,14 @@ public interface IStack
     /// Add element in stack.
     /// </summary>
     /// <param name="value"> Value to add.</param>
-    public void Push(double value);
+    public void Push(int value);
 
     /// <summary>
     /// Returns and removes an element from the top of the stack.
     /// </summary>
     /// <returns> Top element of the stack. </returns>
     /// <exception cref="InvalidOperationException"> Can't to Pop() from empty stack. </exception>
-    public double Pop();
+    public int Pop();
 
     /// <summary>
     /// Checking for stack empty.
